requirements.txt
pandas>=1.0
numpy
matplotlib
scikit-learn
pyyaml
seaborn
jupyter
pytest
